re-organize. ignore redux persist
