+++ actions.ts
@@ -6,7 +6,7 @@ export const setStartCamera = createAction<boolean>("set_start_camera")
 
 export const setPreviewVisible = createAction<boolean>("set_preview_visible")
 
-export const setCapturedImage = createAction<any>("set_captured_image")
+export const setCapturedImage = createAction<string>("set_captured_image")
 
 export const setCameraType = createAction<CameraType>("set_camera_type")
