+++ src/CameraPreview.tsx
@@ -1,0 +1,81 @@
+import { Text, View, TouchableOpacity, ImageBackground } from 'react-native'
+
+export default function CameraPreview({
+	photo,
+	retakePicture,
+	savePhoto,
+}: any) {
+	console.log('sdsfds', photo)
+	return (
+		<View
+			style={{
+				backgroundColor: 'transparent',
+				flex: 1,
+				width: '100%',
+				height: '100%',
+			}}
+		>
+			<ImageBackground
+				source={{ uri: photo && photo.uri }}
+				style={{
+					flex: 1,
+				}}
+			>
+				<View
+					style={{
+						flex: 1,
+						flexDirection: 'column',
+						padding: 15,
+						justifyContent: 'flex-end',
+					}}
+				>
+					<View
+						style={{
+							flexDirection: 'row',
+							justifyContent: 'space-between',
+						}}
+					>
+						<TouchableOpacity
+							onPress={retakePicture}
+							style={{
+								width: 130,
+								height: 40,
+
+								alignItems: 'center',
+								borderRadius: 4,
+							}}
+						>
+							<Text
+								style={{
+									color: '#fff',
+									fontSize: 20,
+								}}
+							>
+								Re-take
+							</Text>
+						</TouchableOpacity>
+						<TouchableOpacity
+							onPress={savePhoto}
+							style={{
+								width: 130,
+								height: 40,
+
+								alignItems: 'center',
+								borderRadius: 4,
+							}}
+						>
+							<Text
+								style={{
+									color: '#fff',
+									fontSize: 20,
+								}}
+							>
+								save photo
+							</Text>
+						</TouchableOpacity>
+					</View>
+				</View>
+			</ImageBackground>
+		</View>
+	)
+}
